+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Moved Glicko config from `URating` to `UGlickoSettings`
+- `UMatch` class is now `FMatch` struct
 
 ## [0.1.0] - 2025-01-09